Split PresetModal into separate components
